DDPM/requirements.txt
absl-py==1.4.0
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
astunparse==1.6.3
backcall==0.2.0
cachetools==4.2.4
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.7
cmake==3.26.3
comm==0.2.2
contourpy==1.0.7
cycler==0.11.0
debugpy==1.8.1
decorator==5.1.1
einops==0.6.1
executing==2.0.1
filelock==3.12.0
flatbuffers==23.5.9
fonttools==4.39.4
gast==0.3.3
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.2
h5py==2.10.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.29.4
ipython==8.12.3
jax==0.4.10
jedi==0.19.1
Jinja2==3.1.2
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
keras==2.13.1
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==16.0.0
lit==16.0.5
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.7
ml-dtypes==0.1.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==4.2.2
prompt-toolkit==3.0.43
protobuf==4.23.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.18.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
pyzmq==26.0.3
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
stack-data==0.6.3
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
threadpoolctl==3.5.0
torch==2.0.1
torchvision==0.15.2
tornado==6.4
tqdm==4.65.0
traitlets==5.14.3
triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.13
Werkzeug==2.3.4
wrapt==1.14.1
zipp==3.15.0
